[V.1-035] FullStack - Temporarily done checkout, change models, etc.

+++ pien_store-backend/database/migrations/2020_04_02_143347_create_order_details_table.php
@@ -16,7 +16,7 @@ class CreateOrderDetailsTable extends Migration
         Schema::create('order_details', function (Blueprint $table) {
             $table->string('order_id', 50);
             $table->string('prod_id', 50);
-            $table->unsignedInteger('quantity');
+            $table->unsignedSmallInteger('quantity');
             $table->primary(['order_id','prod_id']);
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade')->onUpdate('cascade');
             $table->foreign('prod_id')->references('id')->on('products')->onDelete('cascade')->onUpdate('cascade');
